* using size_t

+++ src/o_list.h
@@ -69,7 +69,7 @@ public:
 
     int SetPos(int ARow, int ACol);
     void FixPos();
-    virtual int GetRowLength(int ARow) { return 0; };
+    virtual size_t GetRowLength(int ARow) { return 0; };
 
     virtual void RescanList();
     virtual void UpdateList();
