+++ src/s_string.h
@@ -1,8 +1,8 @@
 #ifndef __S_STRING_H__
 #define __S_STRING_H__
 
-int UnTabStr(char *dest, int maxlen, const char *source, int slen);
-int UnEscStr(char *dest, int maxlen, const char *source, int slen);
+size_t UnTabStr(char *dest, size_t maxlen, const char *source, size_t slen);
+size_t UnEscStr(char *dest, size_t maxlen, const char *source, size_t slen);
 
 #if !defined(HAVE_STRLCPY)
 size_t strlcpy(char *dst, const char *src, size_t size);
